backend/requirements.txt
asgiref==3.5.2
Django==4.0.5
django-environ==0.9.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
